Manage Listings Page & User authorization

+++ app/Http/Controllers/ListingController.php
@@ -64,6 +64,11 @@ class ListingController extends Controller
     // update Listing Data
     public function update(Request $request, Listing $listing){
 
+        // Make sure Logged in
+        if($listing->user_id != auth()->id()){
+            abort(403, 'Unauthorized Action');
+        }
+
         $formFields=$request->validate([
 
             'title'=> 'required',
@@ -91,7 +96,18 @@ class ListingController extends Controller
 
     // Delete Listing
     public function destroy(Listing $listing){
+
+        // Make sure Logged in
+        if($listing->user_id != auth()->id()){
+            abort(403, 'Unauthorized Action');
+        }
+        
         $listing->delete();
         return redirect('/')->with('message','Listing deleted successfully');
+    }
+
+    // Manage Listings
+    public function manage(){
+        return view('listings.manage',['listings'=>auth()->user()->listings()->get()]);
     }
 }
